Add config option to enable links to Cell Broadcast app.
Add "config_cellBroadcastAppLinks" internal boolean config item,
defaulting to false. For devices with the Cell Broadcast app,
setting this config option to true will add a menu item to the
MMS app to jump to the list of received cell broadcasts, and a
preferences item to Settings to jump to cell broadcast settings.

Bug: 6709985
Change-Id: I9fe1919e6d76ba1dfe1e1381e2c38ecdd5e0a867

Former-commit-id: 43af7e075622a315ffef6028264853c178ff4312