Fix issue #6730064: When turning off Nakasi, it very often...
...turns itself immediately back on.

The ON_AFTER_RELEASE flag is documented to not turn the screen on if
it is currently off.

Unfortunately, it didn't seem to actually do this -- it would just
cause a userActivity() call, which turns on the screen if it is
currently off.

Fix this by adding yet another boolean to that function to tell it
to not poke user activity if the screen is off.  (Yes the number of
booleans on it is now insane, and should be cleaned up after we
get through JB.)

Bug: 6730064
Change-Id: I850dfbc777c7668d08b7d63f42a293e22b878256

Former-commit-id: ea401541c51422ba76f8319927cca0f303c57a4e