Fix issue #6706369: Sometimes the "App info" popup shows...
...even though I don't long press

When you start scrolling from a point between two notification
items, only the first down goes to SwipeHelper.onInterceptTouchEvent(),
and the following events go to SwipeHelper.onTouchEvent().  However
when we call SwipeHelper.onTouchEvent(), we immediately bail if we are
not in the drag state, so we never clear the pending long press event.

Bug: 6706369
Change-Id: Icc46fba62fe7ee334d8d62ac39195d7c3aeff705

Former-commit-id: 7f3b379e995651a6ff76d02dddca87b3acdc157e