Add a blank line to MockRilTest.java
Create a merge conflict on the first change to be submitted for moving
telephony and mms from frameworks/base to frameworks/opt. I'm doing this
by adding a blank line to a file in master that will be deleted in
frameworks/base in jb-dev-plus-aosp.

We can then manually merge all of the changes across the projects and
minimize the time master might be broken.

Change-Id: Ie55db5037dbfd20ea6b4be430f65b93250203a63

Former-commit-id: 9b3465792c9de9a1a793b8f36e7f15c4b449c072