Demoted Lohit-Telugu.ttf in fallback lists.
Lohit-Telugu.ttf provides incorrect glyphs for U+FFBC-10007, masking
fonts later in the fallback chain that provide correct glyphs for
that range. I've temporarily demoted Lohit-Telugu in the fallback
order so that this does not happen, and appended a comment that
explains why. In the future, when we have our own Telugu font, we
can place that with the rest of the Indic fonts in the fallback list.

Bug: 6723057
Change-Id: I15fb06dcb7dd13289e226de01c36058134a1b7cb

Former-commit-id: 2a41a2d4ccd2f2037a8f44232486b0e08ed3cd4c