Fix detection of CDMA cell broadcast messages.
The bIsServicePresent field of an incoming CDMA SMS RIL message
is not a reliable way to distinguish broadcast from point-to-point
messages on some RILs. Change CDMA SmsMessage.getMessageType() to
return MESSAGE_TYPE_BROADCAST when the service category is not 0,
and MESSAGE_TYPE_POINT_TO_POINT when the service category is 0.

Bug: 6853691
Change-Id: I0d44a8ce0bb295da5125271db90652570d655bc1

Former-commit-id: 15c257eb8f077fc4f6c95f8a7bc84dd9a7b39c61