Fix shift/mask error in ArtbEvaluator
Shifting from the left copies the MSB along with it. This causes a problem
in ArgbEvaluator, which shifts the top byte down by 24 for the start/end
colors, and then uses those values to interpolate alpha values. The correct
appraoch (used with the other color components) is to mask by 0xff after the
shift.

Issue #6960514 External bug: ArgbEvaluator can't evaluate alpha value properly

Change-Id: I750d38ddfecc5f30d8dab7c6d27d1a7ac06361c3

Former-commit-id: 9b5599894b80b2707909e8e6872eeec7c58af73a