Fix alpha value for device orientation
This seems to have been incorrect since this code was first written for HC. I
can only think that the error was made due to confusion between portrait and
landscape mode as development switched from phones to tablets.

Tested on stingray and prime.

A similar fix will be required for Chrome.

Change-Id: I5da13c489fdb9de340b9e6ea8868c9f1adc15bb2

Former-commit-id: d5be055f79615e2b80fb3fe92a35db2bc84c023d