Use correct WindowManager for Toast.
The WindowManager retrieved from the activity Context with
getSystemService is incorrect for a Toast. Because it contains
a parent window when addView is called its LayoutParams.token
is set to the value of the parent window. Then when an Activity
is dismissed WindowManagerGlobal.closeAll() sees the incorrect
token and incorrectly closes the Toast.

This fix uses the application Context instead of the activity
Context to retrieve a WindowManager with no parent window. This
leaves the token unchanged and keeps from dismissing the toast when
the activity is closed.

Fixes bug 7048792.

Change-Id: I92c3095d8fabd6e9e4206e9bc8e917885ab322a9

Former-commit-id: b6ee2a24292120f95bb5da4539a0672280544be3