Expand the navbar dead zone briefly after touches elsewhere.
For now, each time the dead zone captures a touch, it will
flash to let you know that's what happened. We should
probably turn this off before shipping.

Note also that this currently only expands the navbar on
ACTION_DOWN elsewhere in the UI (because this is the only
time ACTION_OUTSIDE is generated); this isn't perfect and
we'll need to do some mucking about with the input system to
get information about ongoing drags or additional pointers
down, but this CL is a good start and introduces the general
mechanism for expanding and contracting the dead zone.

Change-Id: I43e63aed1e541fd07d83fde4d66bcb5be89b69d2

Former-commit-id: c26185bb7b30508a1940eb452af64b6d8661915f