MediaFormat: Explain that KEY_COLOR_FORMAT is used with decoders as well
If reading the decoded pixel data (instead of letting the system
render it), this is critical for interpreting the data correctly.

The video MediaCodec CTS tests use this MediaFormat key in this
way as well.

Therefore, make sure the documentation reflects the actual use of
this field - claiming that it is encoder-only is misleading.

Change-Id: I9bfd295a95dd3e7d1fa3189886fd243b07bb82cc

Former-commit-id: 158754b1bfc9e0a03a6d9bcd999bdcbd42cedd30