Add getter for animateFirstView property of ViewAnimator
Tooling for xml properties is easier/better if we can properly reflect the
xml properties at the Java language API level as well. We had setters/getters
for the other properties of ViewAnimator, but only a setter for the
animateFirstView property.

Issue #6104327 Inconsistency between XML API and Java API for widgets - ViewAnimator

Change-Id: Iede7231fa433ce14f7a8299da4cedd4720370bc8

Former-commit-id: 201501f11e2934f63f9692383ade0620c45febae