SearchWidget: passing APP_DATA to voice intent
SearchView uses SearchManager.APP_DATA to pass data between the origin activity
and the search activity. This works for text searches (keyboard), but it's not
implemented by voice recognition searches. The method createVoiceAppSearchIntent
initializes a empty bundle that is filled with the QUERY parameter.
This patch adds the SearchManager.APP_DATA parameter if is present, let the
search activity access to the data passed by the origin activity

Patch 2: Fix code style

Change-Id: I52dce34e1efd07a30ad74d52bfc60fbae1c4310b

Former-commit-id: 1bcfe84179417cda57176fe8661a4a749cb843b6