Add Roboto Thin and Thin Italic fonts, as sans-serif-thin family
This change adds the Roboto Thin and Thin Italic font variants, which
are accessible through, for example, Typeface.create("sans-serif-thin",
Typeface.NORMAL).

The version checked in is Roboto_1.100150 as obtained from Christian.
The other fonts in the family should probably be rev'ed as well, but
that would be a separate change.

I verified that the fonts work, that they're thin, that kerning and
ligatures work, and that the "kern" table (redundant and bulky) is
missing.

This should close bug 7173069: Integrate roboto thin to support jb mr1
time displays

Change-Id: I5c950a1843675668b83c005234b44abbe924d9fe

Former-commit-id: be456f2e81ac9c205178883b6cbf880304665319