Don't load views before onResume called.
Views were being kicked off when RecentsTaskLoader was being started
from BaseStatusBar.toggleRecentsActivity. Then RecentTasksLoader called
RecentsPanelView.onTasksLoaded which invalidated the Adapter and loaded
the item views. All this before window manager had determined that the
Activity should be rotated and providing a new Configuration.

This fix waits for the Activity to resume before allowing the Adapter
to be invalidated.

Fixes bug 7138698.

Change-Id: I0df67ff2e07a0e0b69cc3b52e9843f90ce763704

Former-commit-id: ca6da00f6c5381ce21b45bcea4381e8b33f84252